Inclui regras de negocio

+++ src/main/java/com/carloshsrosa/libraryapi/service/AutorService.java
@@ -1,10 +1,15 @@
 package com.carloshsrosa.libraryapi.service;
 
+import com.carloshsrosa.libraryapi.exceptions.AutorPossuiLivrosException;
+import com.carloshsrosa.libraryapi.exceptions.RegistroDuplicadoException;
 import com.carloshsrosa.libraryapi.model.Autor;
 import com.carloshsrosa.libraryapi.repository.AutorRepository;
+import com.carloshsrosa.libraryapi.validator.AutorPossuiLivrosValidator;
+import com.carloshsrosa.libraryapi.validator.AutorValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
 
@@ -13,7 +18,16 @@ public class AutorService {
     @Autowired
     private AutorRepository repository;
 
-    public Autor salvarAutor(Autor autor){
+    @Autowired
+    private AutorValidator validator;
+
+    @Autowired
+    private AutorPossuiLivrosValidator livrosValidator;
+
+    public Autor salvarAutor(Autor autor) {
+        if (validator.validar(autor)){
+            throw new RegistroDuplicadoException("Autor já existente");
+        }
         return repository.save(autor);
     }
 
@@ -22,11 +36,14 @@ public class AutorService {
     }
 
     public void deletarPorId(UUID uuid) {
+        if (livrosValidator.validar(uuid)){
+            throw new AutorPossuiLivrosException("Não é permitido deletar autor que possui livros");
+        }
         repository.deleteById(uuid);
     }
 
     public List<Autor> consultarPorParametros(String nome, String nacionalidade) {
-        if (nome != null && nacionalidade != null){
+        if (nome != null && nacionalidade != null) {
             return repository.findByNomeContainingIgnoreCaseAndNacionalidadeContainingIgnoreCase(nome, nacionalidade);
         }
         if (nome != null) {
@@ -40,6 +57,9 @@ public class AutorService {
     }
 
     public void atualiza(Autor autor) {
+        if (validator.validar(autor)){
+            throw new RegistroDuplicadoException("Autor já existente");
+        }
         repository.save(autor);
     }
 }
